requirements.txt
Django==5.0.6
psycopg
psycopg-binary
psycopg-pool
# django_rest_framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==6.0.0
